Move database to module, add seeder stub.
